refactor(desktop): 提取 workspace mcp 与 skill 命令
